test helper file rename
